+++ Src/navyG.c
@@ -1487,7 +1487,7 @@ ext_navycmd PARM_1 (int, navie)
 	nunit_stat(i)) {
 
       /* check for overruns */
-      if (x + strlen(stat_info[i].input) > COLS - 8) {
+      if (x + (int)strlen(stat_info[i].input) > COLS - 8) {
 	addstr(",");
 	x = 1;
 	move(y++, x);
